Fix hota for empty frames

+++ trackeval/metrics/hota.py
@@ -55,7 +55,7 @@ class HOTA(_BaseMetric):
             # These are normalised, weighted by the match similarity.
             similarity = data['similarity_scores'][t]
             sim_iou_denom = similarity.sum(0)[np.newaxis, :] + similarity.sum(1)[:, np.newaxis] - similarity
-            sim_iou = np.zeros_like(similarity)
+            sim_iou = np.zeros_like(similarity, dtype=potential_matches_count.dtype)
             sim_iou_mask = sim_iou_denom > 0 + np.finfo('float').eps
             sim_iou[sim_iou_mask] = similarity[sim_iou_mask] / sim_iou_denom[sim_iou_mask]
             potential_matches_count[gt_ids_t[:, np.newaxis].tolist(), tracker_ids_t[np.newaxis, :].tolist()] += sim_iou
